Remove active by default profiles as they conflict
Activate by default profiles will turn off if others turn on.  This is a
problem if we want to try to run in both jdk6 and jdk8 due to findbugs
problem.  That said, there was little reason I could see as to why a
profile existed to start with on the multi module build.  All normal
builds build the entire package.  So if someone really doesn't want that
for some reason they can just comment out the two they don't want but
whats a few more minutes of a build anyways.  So squashed that.